requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
mysql-connector-python==8.0.26
numpy==1.24.3
pandas==2.0.3
pytest==6.2.5
httpx==0.25.0
python-dotenv==1.0.0
pydantic==2.4.2